requirements.txt
fastapi
uvicorn[standard]
python-dotenv
crewai
masumi
pydantic
python-multipart
httpx
PyPDF2
litellm